Split remaining heavy spec files for parallel execution
Recorded per-file runtimes show outdated_spec.rb, newgem_spec.rb and
update_spec.rb each exceed the ideal per-worker load of a parallel
run, and setup_spec.rb dominates standalone runs, so file-based
distribution cannot go faster without splitting them. Blocks move
verbatim and example sets are unchanged. newgem's shared examples move
together with their consumers, with its top-level helpers duplicated
into both files.
