+++ spec/support/shards.rb
@@ -10,6 +10,7 @@ module Spec
     EXAMPLE_MAPPINGS = {
       shard_a: [
         "spec/runtime/setup_spec.rb",
+        "spec/runtime/setup_gems_spec.rb",
         "spec/commands/install_spec.rb",
         "spec/commands/add_spec.rb",
         "spec/install/gems/compact_index_spec.rb",
@@ -102,6 +103,7 @@ module Spec
       ],
       shard_c: [
         "spec/commands/newgem_spec.rb",
+        "spec/commands/newgem_options_spec.rb",
         "spec/commands/exec_spec.rb",
         "spec/commands/clean_spec.rb",
         "spec/commands/platform_spec.rb",
@@ -149,7 +151,9 @@ module Spec
         "spec/bundler/resolver/cooldown_spec.rb",
         "spec/install/cooldown_spec.rb",
         "spec/commands/outdated_spec.rb",
+        "spec/commands/outdated_filters_spec.rb",
         "spec/commands/update_spec.rb",
+        "spec/commands/update_scenarios_spec.rb",
         "spec/install/deploy_spec.rb",
         "spec/install/gemfile/sources_spec.rb",
         "spec/runtime/self_management_spec.rb",
